converting make-exception to a primitive opcode instead of a foreign function, since foreign functions will always raise any exception they return (hence the double catch in issue #15).
also restoring the original exception handler when an exception is
raised in with-exception-handler, so that exceptions within the
handler itself don't cause an infinite loop.  this may change, as
with-exception-handler is meant to be a low-level tool on which to
build either guard or condition-case, but until then the restoring
is necessary.