chibi.io: 'feed-pipe' primitive for pipe-like transfer
A separate primitive is necessary to properly pipe data between actual
OS pipes and Scheme ports. We cannot directly connect them and let the
scheduler do the transfer work. Also, IPC interaction via pipes assumes
that the data is simultaneously read from the read end and written to
the write end of our pipe connection, so we cannot use the provided
blocking API to do this as it will result into a deadlock when the read
end is temporarily exhausted or the write end is full.