chibi.process: 'process-pipe' procedure for pipe-like io
feed-pipe requires the pipes to be non-blocking, even if Chibi does not
include support for green threads, so we need to enforce the O_NONBLOCK
flag on them.

Also note that neither of {in,out,err}-port is closed by process-pipe.
It could be reasonable to do this for the in-port, but out- and err-port
must remain open to allow collecting the output from 'virtual ports'
opened by open-input-{string,bytevector}.