sexp: make sexp_write_char() return putc()-like value
sexp_read_char() consistently follows getc() convention regarding the
return value: the character read in case of success, and EOF otherwise.
It is the same for string streams, C file streams, or buffered streams.

sexp_write_char() was defined inconsistently, in some cases it behaved
like putc()--the character written in case of success, EOF otherwise--
while in other cases it returned 0 for success and -1 for failure.

Most of sexp_write_char() usages do not care about the returned value,
but some do and expect putc() convention. This may cause `write-char`
and `write-u8` to not fail when trying to write into a suddenly closed
port, and will be important for proper support of pipes.