chibi.crypto: native SHA-2 implementation
The original Scheme implementation is astonishingly slow. Rewriting
SHA-2 in C yields around x10000 speed boost. This implementation is
on par in terms of speed with the standard sha256sum tool and GPG.

The tricky part is how to get exact 32-bit integers in C89. There
is no <inttypes.h> so we have to try out <limits.h> and see whether
we have a standard type with suitable boundaries.

I have also added some documentation for the exported procedures to
clearly state their input domain.

The intent is to provide one crypto$(SO) library for all native code
in (chibi crypto), so the future MD5 implementation can be placed
into some md5.c and just included into crypto.stub.